requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
openai
unidiff
pytest
anthropic
PyGithub
unidiff
transformers
torch
accelerate
fastmcp